requirements.txt
openai-whisper
torch
pyaudio
pynput
langchain-ollama
langchain-community
langchain-text-splitters
pypdf
chromadb

